Refactor: 주문 메뉴 서비스의 updateOrderTable메서드 수정 - cascade로 인한 주문 추가 및 삭제 처리 - Menu엔티티 속 MenuStatus 상태값을 추가하여 품절 여부 확인 및 salesToday 수량 조절 메서드 추가 Refs: #84

+++ src/main/java/com/beyond/jellyorder/domain/menu/dto/MenuCreateReqDto.java
@@ -37,7 +37,7 @@ public class MenuCreateReqDto {
     private String origin;
 
     @Min(value = -1, message = "판매 한도는 -1 이상이어야 합니다.")
-    private Long salesLimit = -1L;
+    private Integer salesLimit = -1;
 
     @Builder.Default
     private List<String> ingredients = new ArrayList<>();
